+++ ladcore/hook_test.go
@@ -24,7 +24,7 @@ import (
 	"testing"
 
 	. "github.com/tnngo/lad/ladcore"
-	"github.com/tnngo/lad/zaptest/observer"
+	"github.com/tnngo/lad/ladtest/observer"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"