+++ ladcore/sampler_test.go
@@ -29,7 +29,7 @@ import (
 
 	"github.com/tnngo/lad/internal/ztest"
 	. "github.com/tnngo/lad/ladcore"
-	"github.com/tnngo/lad/zaptest/observer"
+	"github.com/tnngo/lad/ladtest/observer"
 	"go.uber.org/atomic"
 
 	"github.com/stretchr/testify/assert"