+++ ladtest/observer/observer_test.go
@@ -29,7 +29,7 @@ import (
 
 	zap "github.com/tnngo/lad"
 	"github.com/tnngo/lad/ladcore"
-	. "github.com/tnngo/lad/zaptest/observer"
+	. "github.com/tnngo/lad/ladtest/observer"
 )
 
 func assertEmpty(t testing.TB, logs *ObservedLogs) {