+++ logger_test.go
@@ -29,7 +29,7 @@ import (
 	"github.com/tnngo/lad/internal/exit"
 	"github.com/tnngo/lad/internal/ztest"
 	"github.com/tnngo/lad/ladcore"
-	"github.com/tnngo/lad/zaptest/observer"
+	"github.com/tnngo/lad/ladtest/observer"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"